feat: Refactor login page and add Snackbar notifications
Replaced inline SVG with a new `D12` component for reuse. Integrated `notistack` into the app for Snackbar functionality and simplified loading state handling by relying on `AuthContext`. Updated file structure and import paths for better consistency.
