+++ web-app/package.json
@@ -20,6 +20,7 @@
         "@tanstack/react-query": "^5.67.3",
         "axios": "^1.8.3",
         "next": "15.3.2",
+        "notistack": "^3.0.2",
         "openapi-fetch": "^0.14.0",
         "openapi-react-query": "^0.5.0",
         "react": "^19.0.0",
